Excercises 4.15-4.23

+++ Part4/Blog/controllers/blogs.js
@@ -1,31 +1,58 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
+const middleware = require('../utils/middleware');
 
 blogRouter.get('/', async (request, response) => {
-  const res = await Blog.find({});
+  const res = await Blog.find({}).populate('user', {username: 1, name: 1});
   response.json(res);
 });
 
-blogRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body);
-    if (!blog.title || !blog.url){
-      return response.status(400).end();
-    }
-    if (!blog.likes) {
-      blog.likes = 0;
-    }
-    const savedBlog = await blog.save();
-    response.status(201).json(savedBlog);
+
+
+blogRouter.post('/', middleware.getToken, async (request, response) => {
+  const body = request.body;
+  const user = request.user;
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  });
+
+  if (!blog.title || !blog.url){
+    return response.status(400).end();
+  }
+  if (!blog.likes) {
+    blog.likes = 0;
+  }
+
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+  response.status(201).json(savedBlog);
 });
 
-blogRouter.delete('/:id', async (req, res) => {
+
+
+blogRouter.delete('/:id', middleware.getToken, async (req, res) => {
   const id = req.params.id;
+  const user = req.user;
+  const blog = await Blog.findById(id);
+  if (!user){
+    return res.status(401).json({error: 'User not found.'});
+  }
+  if (!blog){
+    res.status(404).json({ error: 'Blog not found.' });
+  }
+  if (blog.user.toString() !== user.id.toString()){
+    return res.status(401).json({error: 'Incorrect user.'});
+  }
   const deleted = await Blog.findByIdAndDelete(id);
   if (deleted) {
     res.status(204).json(deleted).end(); 
-  }
-  else {
-    res.status(404).send({ error: 'Blog not found.' });
   }
 })
 
@@ -42,5 +69,6 @@ blogRouter.put('/:id', async (req, res) => {
     res.status(404).send({ error: 'Blog not found.' });
   }
 })
+
 
 module.exports = blogRouter;
